requirements.txt
python-dotenv==1.2.1
oauthlib==3.3.1
langgraph==1.0.2
langchain==1.0.5
langchain-groq==1.0.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-api-python-client==2.187.0